+++ packages/storybook-addon-carbon-theme/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.6](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.3.5...@carbon/storybook-addon-theme@0.3.6) (2021-06-17)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+## [0.3.5](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.3.4...@carbon/storybook-addon-theme@0.3.5) (2021-06-09)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+## [0.3.4](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.3.3...@carbon/storybook-addon-theme@0.3.4) (2021-06-07)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
 ## [0.3.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.3.2...@carbon/storybook-addon-theme@0.3.3) (2021-06-04)
 
 **Note:** Version bump only for package @carbon/storybook-addon-theme